backend/pyproject.toml
[project]
name = "drug-target-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "azure-ai-projects>=2.0.0b2",
    "azure-identity>=1.17.0",
    "azure-search-documents>=11.6.0",
    "azure-cosmos>=4.7.0",
    "azure-storage-blob>=12.22.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "python-docx>=1.1.0",
    "fpdf2>=2.8.0",
    "azure-monitor-opentelemetry>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.30.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
